requirements.txt
Django
python-decouple
psycopg
psycopg2-binary
gunicorn
django-allauth
django-anymail[mailgun]